+++ _tools/check_docs.ts
@@ -19,6 +19,7 @@ import {
   type DocNodeModuleDoc,
   type JsDoc,
   type JsDocTagDocRequired,
+  type JsDocTagParam,
   type Location,
   type TsTypeDef,
 } from "@deno/doc";
@@ -72,6 +73,7 @@ const ENTRY_POINTS = [
   "../json/mod.ts",
   "../jsonc/mod.ts",
   "../log/base_handler.ts",
+  "../log/file_handler.ts",
   "../log/warn.ts",
   "../log/critical.ts",
   "../log/debug.ts",
@@ -80,6 +82,7 @@ const ENTRY_POINTS = [
   "../log/console_handler.ts",
   "../log/formatters.ts",
   "../log/get_logger.ts",
+  "../log/logger.ts",
   "../media_types/mod.ts",
   "../msgpack/mod.ts",
   "../net/mod.ts",
@@ -178,6 +181,36 @@ function assertHasReturnTag(document: { jsDoc: JsDoc; location: Location }) {
       tag.doc !== undefined,
       "@return tag must have a description",
       document,
+    );
+  }
+}
+
+/**
+ * Asserts that a @param tag has a corresponding function definition.
+ */
+function assertHasParamDefinition(
+  document: DocNodeWithJsDoc<DocNodeFunction | ClassMethodDef>,
+  param: JsDocTagParam,
+) {
+  const paramDoc = document.functionDef.params.find((paramDoc) => {
+    if (paramDoc.kind === "identifier") {
+      return paramDoc.name === param.name;
+    } else if (paramDoc.kind === "rest" && paramDoc.arg.kind === "identifier") {
+      return paramDoc.arg.name === param.name;
+    } else if (
+      paramDoc.kind === "assign" && paramDoc.left.kind === "identifier"
+    ) {
+      return paramDoc.left.name === param.name;
+    }
+    return false;
+  });
+
+  if (!paramDoc) {
+    diagnostics.push(
+      new DocumentError(
+        `@param ${param.name} must have a corresponding named function parameter definition.`,
+        document,
+      ),
     );
   }
 }
@@ -295,6 +328,7 @@ function assertHasTypeParamTags(
  * Asserts that a function document has:
  * - A `@typeParam` tag for each type parameter.
  * - A {@linkcode https://jsdoc.app/tags-param | @param} tag for each parameter.
+ * - A parameter definition inside the function for each @param tag.
  * - A {@linkcode https://jsdoc.app/tags-returns | @returns} tag.
  * - At least one {@linkcode https://jsdoc.app/tags-example | @example} tag with
  *   a code snippet that executes successfully.
@@ -307,10 +341,24 @@ function assertFunctionDocs(
     if (param.kind === "identifier") {
       assertHasParamTag(document, param.name);
     }
+    if (param.kind === "rest" && param.arg.kind === "identifier") {
+      assertHasParamTag(document, param.arg.name);
+    }
     if (param.kind === "assign" && param.left.kind === "identifier") {
       assertHasParamTag(document, param.left.name);
     }
   }
+
+  const documentedParams = document.jsDoc.tags?.filter((
+    tag,
+  ): tag is JsDocTagParam =>
+    // Filter nested definitions like options.root as it is still documenting options parameter
+    tag.kind === "param" && !tag.name.includes(".")
+  ) ?? [];
+  for (const param of documentedParams) {
+    assertHasParamDefinition(document, param);
+  }
+
   for (const typeParam of document.functionDef.typeParams) {
     assertHasTypeParamTags(document, typeParam.name);
   }