+++ cbor/_map_decoded_stream_test.ts
@@ -2,8 +2,8 @@
 
 import { assert, assertEquals } from "@std/assert";
 import { random } from "./_common_test.ts";
-import { encodeCbor } from "./encode_cbor.ts";
 import { CborMapDecodedStream } from "./_map_decoded_stream.ts";
+import { encodeCbor } from "./encode_cbor.ts";
 import { CborSequenceDecoderStream } from "./sequence_decoder_stream.ts";
 import type { CborMapStreamOutput } from "./types.ts";